+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.3.0
 
+- #14 Add option `resampling_filter`
 - #14 Add new command `image-factory:generate-destination`
 
 # 0.2.5